+++ src/main/java/com/github/jeffmagina/io/SqlDataSource.java
@@ -4,24 +4,23 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
-public class DataSource {
+public class SqlDataSource {
 
 	private static String url;
 	private static String user;
 	private static String password;
 
 	static {
-		url = System.getProperty("database.url", "jdbc:postgresql://18.219.14.128:5432/jeffMagina");
+		url = System.getProperty("database.url", "jdbc:postgresql://18.216.94.221:5432/jeffMagina");
 		user = System.getProperty("database.username", "jeffMagina");
 		password = System.getProperty("database.password", "jeffMagina");
 	}
 
-	private DataSource() {
+	private SqlDataSource() {
 	}
 
 	public static Connection getConnection() throws SQLException {
 		return DriverManager.getConnection(url, user, password);
-
 	}
 
 }