+++ src/main/java/com/github/jeffmagina/restaurant/main/Main.java
@@ -1,36 +1,35 @@
 package com.github.jeffmagina.restaurant.main;
 
 import java.io.File;
+import java.util.ArrayList;
 
+import com.github.jeffmagina.io.ParseInputFileRepo;
 import com.github.jeffmagina.io.IO;
 import com.github.jeffmagina.restaurant.cashier.Cashier;
 import com.github.jeffmagina.restaurant.customerticket.CustomerTicket;
 import com.github.jeffmagina.restaurant.orderhistory.OrderHistory;
 
+
 public class Main {
 	public static void main(String[] args) {
 		Cashier cashier = new Cashier();
-		File orderForm = new File("OrderForm.txt");
-		IO parser = new IO();
-
-		// Parse order
-		String rawCustomerTicket = parser.readFile(orderForm);
-
-		CustomerTicket customerTicket= new CustomerTicket();
-		customerTicket = parser.ParseCustOrderForm(rawCustomerTicket);
-		
-		//Cashier interactions with customer
-		cashier.interaction(customerTicket);
-		
-		//store into file and database
-		cashier.storeOrder();
-		
-		//displays all orders in OrderHistory
 		OrderHistory orderHistory = new OrderHistory();
-		orderHistory.displayAllOrders();
-		
-		System.out.println("done");
-		
+		File orderForm = new File("OrderForm.txt");
 
+		// Parse Order from text file and store into an array of CustomerTickets
+		ArrayList<CustomerTicket> customerTickets = new ParseInputFileRepo(orderForm).readAll();
+
+		//loops over how many Tickets in orderForm.txt
+		for (int i = 0; i < customerTickets.size(); i++) {
+			// cashier interactions with customer
+			CustomerTicket filledCustomerTicket = cashier.interaction(customerTickets.get(i));
+			
+			// stores data into database and txt file
+			orderHistory = new IO().storeOrder(filledCustomerTicket,"history.txt");	
+		}
+		
+		// display what is in order history, which should match database, and txt file
+		orderHistory.displayAllOrders();
+		System.out.println("done");
 	}
 }