+++ src/main/java/com/github/jeffmagina/restaurant/orderhistory/OrderHistory.java
@@ -6,22 +6,24 @@ import com.github.jeffmagina.restaurant.customerticket.CustomerTicket;
 
 public class OrderHistory extends CustomerTicket {
 
-	public static ArrayList<CustomerTicket> customerTickets = new ArrayList<CustomerTicket>();
+	public ArrayList<CustomerTicket> allCustomerTickets = new ArrayList<CustomerTicket>();
 
-	public OrderHistory() {
-	}
-
-	public void storeOrder(CustomerTicket customerTicket) {
-
-		customerTickets.add(customerTicket);
-		// store in database
-		// orderRepo.insertAll(allOrders);
-	}
-
-	public void displayAllOrders() {
-		for (int i = 0; i < customerTickets.size(); i++) {
-			System.out.println(customerTickets.get(i));
+	public void addCustomerTickets (ArrayList<CustomerTicket> customerTickets) {
+		for(int i = 0; i < customerTickets.size(); i++) {
+			allCustomerTickets.add(customerTickets.get(i));
 		}
 	}
 
+	public void displayAllOrders() {
+		for (int i = 0; i < allCustomerTickets.size(); i++) {
+			System.out.println(allCustomerTickets.get(i));
+		}
+	}
+	
+	public CustomerTicket get(int i) {
+	if (i >= allCustomerTickets.size() || i < 0) {
+		throw new IndexOutOfBoundsException("Index: " + i + ", Size " + i);
+	}
+	return allCustomerTickets.get(i);
+	}
 }